原作者留下的一些警告（如有符号和无符号不匹配、未引用的局部变量等），休改掉。

+++ FxLib/CstBuffer.cpp
@@ -90,7 +90,7 @@ bool		CstBuffer::bindBufferRange(int n, IPass* pPass)
 			updateForTarget(m_activeTarget);
 		} else {
 			Pass* pP = static_cast<Pass*>(pPass);
-			for(int i=0; i<m_targets.size(); i++)
+			for(size_t i=0; i<m_targets.size(); i++)
 			{
 				if((m_targets[i].pass == pPass)&&(m_targets[i].passLayerId == pP->getActiveProgramLayer()))
 				{
